requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
pdfplumber
pandas
gradio
python-dotenv